Debounce BOOT0 readout in the hot loop
Connecting a USB cable sometimes introduces noise on the pins of the USB
to UART converter leading to a low BOOT0 pin. In the main loop of TX
this is checked for indication of a backpack update. Depending on timing
this leads to the whole TX hanging since the serial connection is closed
and the device enters passthrough mode.

Introduce a simple software debouncing to ensure the state of the pin
has settled in case the first readout indicates BOOT0 is 0. This does
not introduce additional latency on the default loop but prevents noise
from breaking the main use case.